Finalize Makefile improvement & add .gitattributes
Continue on improvment in 8fda1018c2275a7606b7b73ea3d14312eb994746, now
the `make` command correctly identifies the output (`libwhamp.a` and
`whamp_engine_test`) and does not rebuild these if not necessary.
Output was verified to be the same as with the old Makefile (using
sha1sum once more).

Adding .gitattributes to help control CR/LF of files in the repo (LF
only in offical repo, but automatically set on checkout). This resulted
in change of src/README.md (going from CR/LF to only LF).

And also add a SECURITY.md file for this repo (based on one in
irfu-matlab).