[CI] add job to download artifact & compile mex
This is untested and may fail, but hopefully it should be able to
download the artifact from "build", install Matlab, and compile a mex
file.

If this works, then this compiled mex file can be used in a test using
test_whamp_module.m (currently in irfu-matlab). With that we would have
a full test; from source code, to Matlab interface and verification of
computed outputs for default inputs.