[CI] add tests in +whamp.test_whamp_module.m
(from irfu-matlab). With this the CI jobs are finalized so that they can
run all the steps: from source code compiling shared libs, mex compiling
and then do input and output verifications.

Note: These outputs shall NOT be considered a hindering factor if the
code is corrected(?) in the future when dealing with issues like the
"complex/real RI" issue #2.